requirements/requirements-tests.txt
pytest-cov==2.5.0
pytest-django==2.9.1
pytest-pythonpath==0.7
model-mommy==1.2.6